# Only reset and clear bindings when necessary